Onboarding Fixes
