+++ src/components/SignUp/Views/Styles.js
@@ -1,0 +1,12 @@
+
+export const styles = (theme) => ({
+  gridColumn: {
+    flexDirection: 'column',
+    alignItems: 'stretch',
+    alignContent: 'stretch',
+  },
+  scrollBox: {
+    textAlign: 'left',
+    gap: 8,
+  },
+})
