+++ src/components/SignUp/Views/index.js
@@ -1,0 +1,9 @@
+
+export { default as FinishView } from './Finish';
+export { default as InputPillView } from './InputPill';
+export { default as InputSearchView } from './InputSearch';
+export { default as InputSelectView } from './InputSelect';
+export { default as InputTextView } from './InputText';
+export { default as EmailView } from './Email';
+export { default as WelcomeView } from './Welcome';
+export { default as WindowView } from './Window';
